refactor: 기물의 움직임 구현 코드 중복 제거
기물이 움직이는 방향 검증을 자식 클래스에서 구현하도록 변경하고 중복 코드는 `Piece` 클래스 메서드로 추출함.
