+++ src/main/java/com/seong/chess/pieces/Bishop.java
@@ -1,7 +1,5 @@
 package com.seong.chess.pieces;
 
-import com.seong.chess.Position;
-
 public class Bishop extends Piece {
 
     private static final char REPRESENTATION = 'b';
@@ -25,13 +23,10 @@ public class Bishop extends Piece {
     }
 
     @Override
-    public Position nextPosition(String sourcePosition, Direction direction, int moveCount) {
-        direction.checkBishop();
-        Position position = Position.convert(sourcePosition);
-        if (moveCount == 0) {
-            return position;
+    public void checkPieceCanMove(Direction direction) {
+        if (direction.isDiagonal()) {
+            return;
         }
-        Position nextPosition = new Position(position.col() + direction.col, position.row() + direction.row);
-        return nextPosition(nextPosition.convert(), direction, moveCount - 1);
+        throw new IllegalArgumentException("비숍은 정방향으로 이동할 수 없습니다.");
     }
 }
