+++ src/main/java/com/seong/chess/pieces/Blank.java
@@ -1,7 +1,5 @@
 package com.seong.chess.pieces;
 
-import com.seong.chess.Position;
-
 public class Blank extends Piece {
 
     private static final char REPRESENTATION = '.';
@@ -21,7 +19,7 @@ public class Blank extends Piece {
     }
 
     @Override
-    public Position nextPosition(String sourcePosition, Direction direction, int moveCount) {
-        return null;
+    public void checkPieceCanMove(Direction direction) {
+
     }
 }
